remove parse step and fix sugarss fail
