--- test/integration/sugarss/index.js
@@ -1,5 +1,0 @@
-require('./index.sss');
-
-module.exports = function () {
-  return 2;
-}
